feat: enhance deposition metadata drawer with new tab structure
- Refactor DepositionMetadataDrawer to support tabbed interface
- Add DepositionMetadataTab for displaying general metadata
- Add DepositionMethodSummaryTab for method-specific details
- Improve organization and separation of concerns
