+++ frontend/packages/data-portal/app/components/Deposition/DepositionMetadataDrawer/DepositionMethodSummaryTab.tsx
@@ -1,0 +1,15 @@
+import { AcquisitionMethodsMetadataTable } from './AcquisitionMethodsMetadataTable'
+import { AnnotationsMethodsMetadataTable } from './AnnotationsMethodsMetadataTable'
+import { ExperimentalConditionsMetadataTable } from './ExperimentalConditionsMetadataTable'
+import { TomogramMethodsMetadataTable } from './TomogramMethodsMetadataTable'
+
+export function DepositionMethodSummaryTab() {
+  return (
+    <>
+      <AnnotationsMethodsMetadataTable />
+      <TomogramMethodsMetadataTable />
+      <AcquisitionMethodsMetadataTable />
+      <ExperimentalConditionsMetadataTable />
+    </>
+  )
+}
